requirements.txt
transformers==4.45.2
jinja2==3.1.0
pyarrow==18.0.0
datasets==4.0.0
accelerate
peft
bitsandbytes
torch
huggingface_hub
pillow
qwen-vl-utils
openai
requests
matplotlib
pandas
wandb
httpx
dotenv
pytest
torchvision
opencv-python
